Allow overriding items for each control
With this it's possible to customize specific alarm controls, like using a
different pattern if needed or sending specific alarms to different topics.

The `for_each` condition has been changed and `var.create` is now applied when
generating `local.controls`. This is because Terraform complained that the
expression `var.create ? x : y` was returning different types for the true and
false conditions. By filtering controls early we sidestep the problem.